+++ release_notes.md
@@ -4,7 +4,7 @@ This is the first series-1 release! It is backwards compatible.
 
 ## new features 
 
-- units metadata: all output methods now contain a units structure that contains the units for each output field. 
+- units metadata: all output methods now contain a units structure that contains the units for each output field. You can also call `SV_input_units()` to get a structure that lists the expected units for each `VBR.in.SV` field.
 
 ## bug fixes